+++ packages/zefyr/test/testing.dart
@@ -134,7 +134,7 @@ class _ZefyrSandboxState extends State<_ZefyrSandbox> {
     return new ZefyrEditor(
       controller: widget.controller,
       focusNode: widget.focusNode,
-      enabled: _enabled,
+      mode: _enabled ? ZefyrMode.edit : ZefyrMode.view,
       autofocus: widget.autofocus,
     );
   }